requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
argcomplete==3.6.0
astroid==2.9.3
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
boto==2.49.0
bottle==0.12.23
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cycler==0.11.0
deep-translator==1.8.3
discord.py>=2.3.2
docopt==0.6.2
ec2==0.4.0
Eel==0.14.0
feedparser==6.0.8
firewall==0.2.0
flake8==4.0.1
fonttools==4.37.1
future==0.18.2
get-chefkoch==1.0.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
gTTS==2.2.4
idna==3.3
isort==5.10.1
joblib==1.1.0
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lyricsgenius==3.0.1
markovify==0.9.3
matplotlib==3.5.3
mccabe==0.6.1
multidict==6.0.2
nltk==3.7
numpy==1.23.2
packaging==21.3
pep8==1.7.1
Pillow==9.1.1
pipreqs==0.4.11
platformdirs==2.5.1
praw @ https://github.com/praw-dev/praw/archive/master.zip
prawcore==2.3.0
pycodestyle==2.8.0
pycparser==2.20
pydub==0.25.1
pyflakes==2.4.0
pylast==4.5.0
pylint==2.12.2
PyNaCl==1.5.0
pyparsing==3.0.9
pysbd==0.3.4
python-augeas==1.1.0
python-dateutil==2.8.2
pytz==2021.3
regex==2022.7.25
requests==2.27.1
scipy==1.9.0
sgmllib3k==1.0.0
soupsieve==2.3.1
syllables==1.0.3
textblob==0.17.1
textblob-de==0.4.3
toml==0.10.2
torch==1.12.1
tqdm==4.64.0
typing_extensions==4.1.1
Unidecode==1.3.3
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.3.1
whichcraft==0.6.1
wikipedia==1.4.0
wrapt==1.13.3
yarg==0.1.9
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
